tests for ProducerDaoImpl and CarProducerServiceImpl where implemented

+++ app-core/src/test/java/com/epam/brest/course2015/domain/ProducerTest.java
@@ -19,7 +19,7 @@ public class ProducerTest {
     @Test
     public void testGetProducerId(){
         producer.setProducerId(1);
-        assertEquals(producer.getProducerId(), 1);
+        assertEquals(producer.getProducerId(),(Integer)1);
     }
 
     @Test
